refactor: 🔄 testIsExact: index.ts => is-exact.ts
